refactor : BridgeGame에서 View 기능 안 쓰도록 리팩토링
